requirements.txt
absl-py==0.9.0
aiohttp==3.7.3
aiohttp-cors==0.7.0
aioredis==1.3.1
astor==0.8.1
astunparse==1.6.3
async-timeout==3.0.1
atari-py==0.2.6
attrs==19.3.0
beautifulsoup4==4.9.3
blessings==1.7
boto3==1.16.37
botocore==1.19.37
Box2D==2.3.10
box2d-py==2.3.8
cachetools==4.1.1
certifi==2020.12.5
cffi==1.14.1
chardet==3.0.4
click==7.1.2
cloudpickle==1.3.0
colorama==0.4.4
colorful==0.5.4
Cython==0.29
dm-tree==0.1.5
expiringdict==1.2.1
filelock==3.0.12
Flask==1.1.2
future==0.18.2
gast==0.3.3
gitdb==4.0.5
GitPython==3.1.11
glfw==1.12.0
google==3.0.0
google-api-core==1.22.0
google-auth==1.20.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.52.0
gpustat==0.6.0
grpcio==1.31.0
gym==0.17.2
h11==0.11.0
h5py==2.10.0
hiredis==1.1.0
idna==2.10
imageio==2.9.0
importlib-metadata==1.7.0
iniconfig==1.0.1
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
jsonpatch==1.28
jsonpointer==2.0
jsonschema==3.2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
lockfile==0.12.2
lz4==3.1.0
Markdown==3.2.2
MarkupSafe==1.1.1
more-itertools==8.4.0
msgpack==1.0.0
multidict==5.1.0
numpy==1.18.5
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
opencensus==0.7.11
opencensus-context==0.1.2
opencv-python==4.3.0.38
opencv-python-headless==4.3.0.38
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.0
Pillow==7.2.0
pluggy==0.13.1
prometheus-client==0.9.0
protobuf==3.12.4
psutil==5.7.3
py==1.9.0
py-spy==0.3.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.6.1
pyglet==1.5.11
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==6.0.1
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
ray==1.0.1.post1
redis==3.4.1
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.3.3
scipy==1.4.1
six==1.15.0
smmap==3.0.4
soupsieve==2.0.1
tabulate==0.8.7
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorboardX==2.1
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
toml==0.10.1
typing-extensions==3.7.4.2
urllib3==1.26.2
uvicorn==0.13.0
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.5.1
zipp==3.1.0
